backend-fastapi/pyproject.toml
[project]
name = "abhaydhir-backend"
version = "0.1.0"
description = "FastAPI backend for Abhaydhir application"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "supabase",
    "python-dotenv",
    "pyjwt",
    "pydantic",
    "pillow",
    "qrcode",
    "python-multipart",
    "python-jose[cryptography]"
]

